requirements.txt
Django>=5.0.0,<5.3
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.2.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
supabase>=2.0.0

